+++ CHANGES.rst
@@ -4,6 +4,13 @@ Changes
 2.0 (unreleased)
 ----------------
 
+- Prevent initialize if widget is part of array template.
+  [lenadax]
+
+- Extend JS by ``location_on_array_add`` and ``register_array_subscribers``
+  functions to enable usage in ``yafowil.widget.array``.
+  [lenadax]
+
 - Rewrite JavaScript using ES6.
   [rnix]